refactor: split code doc viewer components
