+++ src/widgets/CodeDocView/ui/Icon.tsx
@@ -1,0 +1,72 @@
+import type React from 'react';
+
+export const Icon = ({ name, className = 'w-4 h-4' }: { name: string; className?: string }) => {
+  const paths: Record<string, React.ReactNode> = {
+    home: <path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />,
+    chevronRight: <path d="m9 18 6-6-6-6" />,
+    package: (
+      <path d="m16.5 9.4-9-5.19M21 16V8a2 2 0 0 0-1-1.73l-7-4a2 2 0 0 0-2 0l-7 4A2 2 0 0 0 3 8v8a2 2 0 0 0 1 1.73l7 4a2 2 0 0 0 2 0l7-4A2 2 0 0 0 21 16z" />
+    ),
+    layers: <path d="M12 2L2 7l10 5 10-5-10-5zM2 17l10 5 10-5M2 12l10 5 10-5" />,
+    function: (
+      <>
+        <path d="M19 3v18H5V3h14z" />
+        <path d="M9 15l3-3 3 3" />
+        <path d="M12 9v6" />
+      </>
+    ),
+    interface: (
+      <>
+        <rect width="18" height="18" x="3" y="3" rx="2" />
+        <path d="M3 9h18" />
+        <path d="M9 21V9" />
+      </>
+    ),
+    class: <path d="M4 19.5v-15A2.5 2.5 0 0 1 6.5 2H20v20H6.5a2.5 2.5 0 0 1 0-5H20" />,
+    branch: (
+      <>
+        <path d="M6 3v12" strokeDasharray="4 4" />
+        <circle cx="6" cy="18" r="3" />
+        <path d="M9 18h9l-3-3m3 3l-3 3" />
+      </>
+    ),
+    tag: (
+      <>
+        <path d="M12 2H2v10l9.29 9.29c.94.94 2.48.94 3.42 0l6.58-6.58c.94-.94.94-2.48 0-3.42L12 2Z" />
+        <path d="M7 7h.01" />
+      </>
+    ),
+    info: <circle cx="12" cy="12" r="10" />,
+    arrowDown: <path d="m6 9 6 6 6-6" />,
+    arrowUp: <path d="m18 15-6-6-6 6" />,
+    flow: <path d="M22 7M2 7l10-5 10 5-10 5-10-5zM2 17l10 5 10-5M2 12l10 5 10-5" />,
+    testSuite: <path d="M12 2L2 7l10 5 10-5-10-5zM2 17l10 5 10-5M2 12l10 5 10-5" />,
+    testCase: (
+      <>
+        <circle cx="12" cy="12" r="10" />
+        <path d="M9 12l2 2 4-4" />
+      </>
+    ),
+    testHook: (
+      <>
+        <circle cx="12" cy="12" r="3" />
+        <path d="M12 1v6m0 6v6M5.6 5.6l4.2 4.2m4.2 4.2l4.2 4.2M1 12h6m6 0h6M5.6 18.4l4.2-4.2m4.2-4.2l4.2-4.2" />
+      </>
+    ),
+  };
+
+  return (
+    <svg
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className={className}
+    >
+      {paths[name] || <circle cx="12" cy="12" r="10" />}
+    </svg>
+  );
+};
